requirements.txt
Click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
pymongo
Werkzeug
flask-mail
Flask-Limiter
Flask-Cors
flask-wtf
Pillow